♻️ Refactor: 로컬 일정을 서버 일정으로 변환하는 코드 리펙토링
Fragment에 있던 코드를 Data Class의 메서드로 이동

Related to: #134
